Move shapeless-based builders/refinement to it's own sub-project and package
