+++ checklist/src/main/scala/checklist/syntax.scala
@@ -1,3 +1,3 @@
 package checklist
 
-object syntax extends Rule1Syntax with RuleHListSyntax with MessageSyntax with CheckedSyntax
+object syntax extends Rule1Syntax with MessageSyntax with CheckedSyntax
