+++ refinement/src/main/scala/checklist/refinement/syntax.scala
@@ -1,0 +1,3 @@
+package checklist.refinement
+
+object syntax extends RuleHListSyntax
